+++ libs/classifier/cnb_estimator.h
@@ -1,4 +1,4 @@
-double **CNB_get_param_vecs(
+float **CNB_get_param_vecs(
         int num_classes,
         int num_tokens,
         int *class_samplesizes,
